The append windows move to stmts_append.rs, keeping stmts.rs under the file budget
